+++ ultragrid/src/video_display/decklink.h
@@ -57,7 +57,7 @@ struct video_frame  *display_decklink_getf(void *state);
 int                  display_decklink_putf(void *state, char *frame);
 void                 display_decklink_reconfigure(void *state,
                                 struct video_desc desc);
-int                  display_decklink_get_property(void *state, int property, void *val, int *len);
+int                  display_decklink_get_property(void *state, int property, void *val, size_t *len);
 
 struct audio_frame * display_decklink_get_audio_frame(void *state);
 void 		 display_decklink_put_audio_frame(void *state, struct audio_frame *frame);
