+++ ultragrid/src/video_display/dvs.h
@@ -63,7 +63,7 @@ struct video_frame  *display_dvs_getf(void *state);
 int                  display_dvs_putf(void *state, char *frame);
 void                 display_dvs_reconfigure(void *state,
                                 struct video_desc desc);
-int                  display_dvs_get_property(void *state, int property, void *val, int *len);
+int                  display_dvs_get_property(void *state, int property, void *val, size_t *len);
 
 struct audio_frame * display_dvs_get_audio_frame(void *state);
 void display_dvs_put_audio_frame(void *state, struct audio_frame *frame);
