+++ ultragrid/src/video_display/sdl.h
@@ -57,7 +57,7 @@ struct video_frame *display_sdl_getf (void *state);
 int display_sdl_putf (void *state, char *frame);
 
 void display_sdl_reconfigure(void *state, struct video_desc desc);
-int display_sdl_get_property(void *state, int property, void *val, int *len);
+int display_sdl_get_property(void *state, int property, void *val, size_t *len);
 
 struct audio_frame * display_sdl_get_audio_frame(void *state);
 void display_sdl_put_audio_frame(void *state, const struct audio_frame *frame);
